# api/openapi.yaml
openapi: 3.1.0
info:
  title: K8s Service Provider
  description: A microservice for managing container and virtual machine deployments
  version: 1.0.0
  contact:
    name: K8s Team
    email: k8s-team@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://k8s-service-provider.example.com/api/v1
    description: Production server

paths:
  /deployments:
    post:
      summary: Create a new deployment
      description: Creates a new container or virtual machine deployment
      operationId: CreateDeployment
      parameters:
        - name: id
          in: query
          description: Optional client-provided deployment ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deployment'
            examples:
              container:
                summary: Container deployment example
                value:
                  kind: container
                  metadata:
                    name: nginx-app
                    namespace: default
                    labels:
                      app: nginx
                      version: "1.0"
                  spec:
                    container:
                      image: nginx:latest
                      replicas: 3
                      ports:
                        - container_port: 80
                          service_port: 8080
                      resources:
                        cpu: 100m
                        memory: 128Mi
              vm:
                summary: Virtual machine deployment example
                value:
                  kind: vm
                  metadata:
                    name: fedora-vm
                    namespace: default
                    labels:
                      os: fedora
                      environment: development
                  spec:
                    vm:
                      ram: 4
                      cpu: 2
                      os: fedora
              vm-with-ssh:
                summary: Virtual machine with SSH key example
                value:
                  kind: vm
                  metadata:
                    name: ubuntu-vm
                    namespace: default
                    labels:
                      os: ubuntu
                      environment: production
                  spec:
                    vm:
                      ram: 8
                      cpu: 4
                      os: ubuntu
                      ssh_public_key: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC3F... user@hostname"
                      ssh_key_name: "ubuntu-vm-ssh-key"
      responses:
        '201':
          description: Deployment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Deployment already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List all deployments
      description: Retrieves a list of all deployments
      operationId: ListDeployments
      parameters:
        - name: namespace
          in: query
          description: Filter deployments by namespace
          schema:
            type: string
        - name: kind
          in: query
          description: Filter deployments by kind (container or vm)
          schema:
            type: string
            enum: [container, vm]
        - name: max_page_size
          in: query
          description: Maximum number of deployments to return per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: page_token
          in: query
          description: Token for retrieving the next page of results
          schema:
            type: string
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'
                  next_page_token:
                    type: string
                    description: Token for the next page of results
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /deployments/{deployment}:
    get:
      summary: Get a deployment by ID
      description: Retrieves a specific deployment by its ID
      operationId: GetDeployment
      parameters:
        - name: deployment
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update a deployment
      description: Updates an existing deployment using merge-patch semantics
      operationId: UpdateDeployment
      parameters:
        - name: deployment
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/Deployment'
      responses:
        '200':
          description: Deployment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a deployment
      description: Deletes a specific deployment
      operationId: DeleteDeployment
      parameters:
        - name: deployment
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
      responses:
        '204':
          description: Deployment deleted successfully
        '404':
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      summary: Health check
      description: Returns the health status of the service
      operationId: HealthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  check_time:
                    type: string
                    format: date-time

components:
  schemas:
    Deployment:
      type: object
      x-aep-resource:
        type: "k8s-service-provider/deployment"
        singular: "deployment"
        plural: "deployments"
        patterns:
          - "deployments/{deployment}"
      required:
        - kind
        - metadata
        - spec
      properties:
        path:
          type: string
          readOnly: true
          description: "Resource path. Format: deployments/{deployment}"
          example: "deployments/abc123"
        id:
          type: string
          readOnly: true
          description: Unique deployment identifier
        kind:
          type: string
          enum: [container, vm]
          description: Type of deployment (container or virtual machine)
        metadata:
          $ref: '#/components/schemas/Metadata'
        spec:
          oneOf:
            - $ref: '#/components/schemas/ContainerSpec'
            - $ref: '#/components/schemas/VMSpec'
          discriminator:
            propertyName: kind
            mapping:
              container: '#/components/schemas/ContainerSpec'
              vm: '#/components/schemas/VMSpec'
        status:
          $ref: '#/components/schemas/DeploymentStatus'
          readOnly: true
        create_time:
          type: string
          format: date-time
          readOnly: true
        update_time:
          type: string
          format: date-time
          readOnly: true

    Metadata:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
          description: Deployment name (DNS-1123 compliant)
        namespace:
          type: string
          default: default
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
          description: Kubernetes namespace
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value pairs for labeling

    ContainerSpec:
      type: object
      required:
        - container
      properties:
        container:
          type: object
          required:
            - image
          properties:
            image:
              type: string
              description: Container image name and tag
            replicas:
              type: integer
              minimum: 1
              maximum: 100
              default: 1
              description: Number of replicas
            ports:
              type: array
              items:
                type: object
                required:
                  - container_port
                properties:
                  container_port:
                    type: integer
                    minimum: 1
                    maximum: 65535
                  service_port:
                    type: integer
                    minimum: 1
                    maximum: 65535
                  protocol:
                    type: string
                    enum: [TCP, UDP]
                    default: TCP
            resources:
              type: object
              properties:
                cpu:
                  type: string
                  pattern: '^[0-9]+m?$'
                  description: CPU request (e.g., 100m, 1)
                memory:
                  type: string
                  pattern: '^[0-9]+[KMGT]?i?$'
                  description: Memory request (e.g., 128Mi, 1Gi)
            environment:
              type: array
              items:
                type: object
                required:
                  - name
                  - value
                properties:
                  name:
                    type: string
                  value:
                    type: string

    VMSpec:
      type: object
      required:
        - vm
      properties:
        vm:
          type: object
          required:
            - ram
            - cpu
            - os
          properties:
            ram:
              type: integer
              minimum: 1
              maximum: 32
              description: Memory allocation in GB
            cpu:
              type: integer
              minimum: 1
              maximum: 32
              description: Number of CPU cores
            os:
              type: string
              enum: [fedora, ubuntu, centos, rhel]
              description: Operating system for the VM
            ssh_public_key:
              type: string
              description: |
                Optional SSH public key content to inject into the VM.
                Must be a valid SSH public key (ssh-rsa, ssh-ed25519, ecdsa-sha2-*, or ssh-dss).
                If provided without ssh_key_name, a secret with a randomized name will be created.
                If ssh_key_name is provided and the secret exists, this field is ignored.
              example: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC..."
            ssh_key_name:
              type: string
              pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
              maxLength: 253
              description: |
                Optional name of the Kubernetes Secret to store/reference the SSH public key.
                Must follow DNS-1123 subdomain rules (lowercase alphanumeric with hyphens).
                If not provided and ssh_public_key is set, a randomized secret name will be generated.
                If provided and the secret exists, it will be reused (ssh_public_key is ignored).
                If provided and the secret doesn't exist, ssh_public_key must be provided to create it.
              example: "my-vm-ssh-key"

    DeploymentStatus:
      type: object
      properties:
        phase:
          type: string
          enum: [pending, running, succeeded, failed, unknown]
        message:
          type: string
          description: Human-readable status message
        ready_replicas:
          type: integer
          description: Number of ready replicas (for containers)
        conditions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              status:
                type: string
                enum: ['True', 'False', Unknown]
              last_transition_time:
                type: string
                format: date-time
              reason:
                type: string
              message:
                type: string

    Error:
      type: object
      required:
        - code
        - message
      properties:
        type:
          type: string
          format: uri-reference
          description: Error type identifier
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        error_time:
          type: string
          format: date-time

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []